requirements.txt
Flask==2.3.3             # Core web framework
Flask-SQLAlchemy==3.0.5  # SQLAlchemy integration with Flask
Flask-WTF==1.1.1         # WTForms integration with Flask (for forms)
Flask-Migrate==4.0.4     # Handle database migrations
Flask-Login==0.6.3       # User session management
Flask-Mail==0.9.1        # Email support in Flask
Flask-RESTful==0.3.9     # Building RESTful APIs with Flask
Flask-Cors==4.0.2        # Enable Cross-Origin Resource Sharing (CORS)
Flask-Bcrypt==1.0.1      # Password hashing utilities
Flask-Bootstrap==3.3.7.1 # Bootstrap integration for Flask (if using Bootstrap for UI)
Jinja2==3.1.2            # Template engine used by Flask
SQLAlchemy==2.0.23       # Database ORM
WTForms==3.0.1           # Form handling
email-validator==1.3.1   # Email validation for forms
python-dotenv==1.0.0     # Manage environment variables
gunicorn==21.0.1         # WSGI HTTP Server for deploymentFlask-PyMongo==2.3.0
pymongo==4.3.3
Flask-PyMongo==2.3.0
werkzeug>=2.3.7
Flask-Session==0.4.0